debug capsicum

+++ queues/posix.c
@@ -95,10 +95,10 @@ queue_err_t queue_open(void *p, const char *filename, int flags)
         return QUEUE_ERR_GENERAL_FAILURE;
     }
     if (!HAS_FLAG(flags, QUEUE_FL_SENDER)) {
-        CAP_RIGHTS_LIMIT(q->mq, CAP_READ);
+        CAP_RIGHTS_LIMIT((int) q->mq, CAP_READ);
 #if 0
     } else {
-        CAP_RIGHTS_LIMIT(q->mq, CAP_WRITE);
+        CAP_RIGHTS_LIMIT((int) q->mq, CAP_WRITE);
 #endif
     }
     if (HAS_FLAG(flags, QUEUE_FL_OWNER)) {
